requirements_dev.txt
# These packages will greatly aid development efforts, but should *not* be
# installed in a production environment!

django-debug-toolbar==0.9.4
Werkzeug==0.6.2
django-extensions==0.5
pep8==0.6.1
pyflakes==0.5.0
